# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: CoachLink Platform API
  description: |
    API для платформы автоматизации тренировочного процесса
    и взаимодействия в системе «тренер — спортсмен».

    ## Аутентификация
    Все запросы (кроме /auth/*) требуют JWT access-token
    в заголовке `Authorization: Bearer <token>`.

    ## Роли
    - `coach` — тренер
    - `athlete` — спортсмен

    ## Формат дат
    - Даты в формате ISO 8601: `2026-03-28`
    - Timestamps в формате ISO 8601 с таймзоной: `2026-03-28T12:00:00Z`

    ## Пагинация
    Списочные эндпоинты поддерживают query-параметры `page` и `page_size`.
    Ответ содержит объект `pagination` с метаданными.
  version: 1.0.0
  contact:
    name: CoachLink Team
servers:
  - url: http://localhost:8080
    description: Local development (API Gateway)
...
paths:
...
  /api/v1/users/me:
    get:
      tags: [Users]
      summary: Получить профиль текущего пользователя
      operationId: getMyProfile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Профиль пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/users/search:
    get:
      tags: [Users]
      summary: Поиск пользователей по ФИО или логину
      description: |
        Используется для:
        - Поиск тренера (спортсмен ищет тренера для привязки)
        - Поиск спортсмена (тренер ищет спортсмена для добавления в группу)

        Поиск работает по подстроке (автодополнение).
        Минимальная длина запроса: 2 символа.
      operationId: searchUsers
      security:
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          description: Строка поиска (ФИО или логин), минимум 2 символа
          schema:
            type: string
            minLength: 2
            example: "Иванов"
        - name: role
          in: query
          required: false
          description: Фильтр по роли
          schema:
            type: string
            enum: [coach, athlete]
            example: coach
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Результаты поиска
          content:
            application/json:
              schema:
                type: object
                required: [items, pagination]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserProfile'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ----------------------------------------------------------
  # CONNECTIONS
  # ----------------------------------------------------------
  /api/v1/connections/request:
    post:
      tags: [Connections]
      summary: Отправить заявку тренеру
      description: |
        Спортсмен отправляет заявку на привязку к тренеру.
        У спортсмена может быть только один тренер.
        Если спортсмен уже привязан к тренеру, запрос вернёт ошибку.
      operationId: sendConnectionRequest
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectionRequestCreate'
      responses:
        '201':
          description: Заявка отправлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionRequest'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Только спортсмен может отправлять заявки
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Заявка уже существует или спортсмен уже привязан к тренеру
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                alreadyHasCoach:
                  value:
                    error:
                      code: ALREADY_HAS_COACH
                      message: "Athlete is already connected to a coach"
                requestExists:
                  value:
                    error:
                      code: REQUEST_ALREADY_EXISTS
                      message: "A pending request to this coach already exists"
  /api/v1/connections/requests/incoming:
    get:
      tags: [Connections]
      summary: Получить входящие заявки (для тренера)
      operationId: getIncomingRequests
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          description: Фильтр по статусу
          schema:
            type: string
            enum: [pending, accepted, rejected]
            default: pending
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Список заявок
          content:
            application/json:
              schema:
                type: object
                required: [items, pagination]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ConnectionRequest'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Доступно только тренерам
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/connections/requests/outgoing:
    get:
      tags: [Connections]
      summary: Получить исходящую заявку (для спортсмена)
      description: Возвращает текущую заявку спортсмена (pending), если она есть.
      operationId: getOutgoingRequest
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Текущая заявка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Доступно только спортсменам
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Нет активной заявки
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/connections/requests/{requestId}/accept:
    put:
      tags: [Connections]
      summary: Принять заявку спортсмена
      operationId: acceptConnectionRequest
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: Заявка принята
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Доступно только тренерам / это не ваша заявка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/connections/requests/{requestId}/reject:
    put:
      tags: [Connections]
      summary: Отклонить заявку спортсмена
      operationId: rejectConnectionRequest
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: Заявка отклонена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Доступно только тренерам / это не ваша заявка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/connections/athletes:
    get:
      tags: [Connections]
      summary: Список спортсменов тренера
      description: Возвращает список спортсменов, привязанных к текущему тренеру.
      operationId: getCoachAthletes
      security:
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          required: false
          description: Поиск по ФИО или логину среди своих спортсменов
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Список спортсменов
          content:
            application/json:
              schema:
                type: object
                required: [items, pagination]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/AthleteInfo'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Доступно только тренерам
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/connections/coach:
    get:
      tags: [Connections]
      summary: Получить информацию о тренере спортсмена
      operationId: getAthleteCoach
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Информация о тренере
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CoachInfo'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Доступно только спортсменам
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Спортсмен не привязан к тренеру
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/connections/athletes/{athleteId}:
    delete:
      tags: [Connections]
      summary: Отвязать спортсмена
      operationId: removeAthlete
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AthleteId'
      responses:
        '204':
          description: Спортсмен отвязан
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Доступно только тренерам / это не ваш спортсмен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ----------------------------------------------------------
  # GROUPS
  # ----------------------------------------------------------
  /api/v1/groups:
    post:
      tags: [Groups]
      summary: Создать тренировочную группу
      operationId: createGroup
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupRequest'
      responses:
        '201':
          description: Группа создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainingGroup'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Доступно только тренерам
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags: [Groups]
      summary: Список тренировочных групп
      description: |
        - Тренер видит свои группы.
        - Спортсмен видит группы, в которых он состоит.
      operationId: getGroups
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Список групп
          content:
            application/json:
              schema:
                type: object
                required: [items, pagination]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/TrainingGroupSummary'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/groups/{groupId}:
    get:
      tags: [Groups]
      summary: Детали группы с составом
      operationId: getGroup
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupId'
        - name: q
          in: query
          required: false
          description: Поиск по ФИО или логину внутри группы
          schema:
            type: string
      responses:
        '200':
          description: Детали группы
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainingGroupDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Нет доступа к этой группе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Groups]
      summary: Обновить название группы
      operationId: updateGroup
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGroupRequest'
      responses:
        '200':
          description: Группа обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainingGroup'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Доступно только тренерам / это не ваша группа
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Groups]
      summary: Удалить группу
      operationId: deleteGroup
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupId'
      responses:
        '204':
          description: Группа удалена
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Доступно только тренерам / это не ваша группа
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/groups/{groupId}/members:
    post:
      tags: [Groups]
      summary: Добавить спортсмена в группу
      operationId: addGroupMember
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddGroupMemberRequest'
      responses:
        '201':
          description: Спортсмен добавлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupMember'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Доступно только тренерам / это не ваша группа
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Спортсмен уже в группе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/groups/{groupId}/members/{athleteId}:
    delete:
      tags: [Groups]
      summary: Удалить спортсмена из группы
      operationId: removeGroupMember
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupId'
        - $ref: '#/components/parameters/AthleteId'
      responses:
        '204':
          description: Спортсмен удалён из группы
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Доступно только тренерам / это не ваша группа
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ----------------------------------------------------------
  # INTERNAL API (service-to-service, not for mobile clients)
  # ----------------------------------------------------------
  /internal/users/{userId}:
    get:
      tags: [Internal]
      summary: '[Internal] Получить профиль пользователя по ID'
      description: |
        **Внутренний эндпоинт для межсервисного взаимодействия.**
        Используется Training Service для получения ФИО и логина спортсмена
        при создании тренировочных планов с индивидуальными athlete_ids.
        Не требует JWT-авторизации. Не предназначен для мобильного клиента.
      operationId: internalGetUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Профиль пользователя (упрощённый)
          content:
            application/json:
              schema:
                type: object
                properties:
                  athlete_id:
                    type: string
                    format: uuid
                  full_name:
                    type: string
                  login:
                    type: string
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access-token, полученный при /auth/login или /auth/refresh

  # ----------------------------------------------------------
  # Parameters
  # ----------------------------------------------------------
  parameters:
    RequestId:
      name: requestId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    AthleteId:
      name: athleteId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Unauthorized:
      description: Не авторизован — невалидный или отсутствующий токен
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UNAUTHORIZED
              message: "Invalid or expired access token"
...
  schemas:
...
    UserProfile:
      type: object
      required: [id, login, email, full_name, role, created_at]
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        login:
          type: string
          example: "ivan-petrov"
        email:
          type: string
          format: email
          example: "ivan@example.com"
        full_name:
          type: string
          example: "Петров Иван Сергеевич"
        role:
          type: string
          enum: [coach, athlete]
          example: "athlete"
        created_at:
          type: string
          format: date-time
          example: "2026-03-28T12:00:00Z"
...
    ConnectionRequestCreate:
      type: object
      required: [coach_id]
      properties:
        coach_id:
          type: string
          format: uuid
          description: ID тренера, которому отправляется заявка
...